requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
flask-restplus==0.12.1
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
PyJWT==1.6.4
pytest==3.9.1
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1